+++ packages/auth-construct/src/construct.test.ts
@@ -13,10 +13,10 @@ import {
   CfnIdentityPool,
   CfnUserPool,
   CfnUserPoolClient,
+  OAuthScope,
   UserPool,
   UserPoolClient,
   UserPoolIdentityProviderSamlMetadataType,
-  VerificationEmailStyle,
 } from 'aws-cdk-lib/aws-cognito';
 import { authOutputKey } from '@aws-amplify/backend-output-schemas';
 import { Code, Function, Runtime } from 'aws-cdk-lib/aws-lambda';
@@ -98,7 +98,7 @@ void describe('Auth construct', () => {
   void it('creates phone number login mechanism', () => {
     const app = new App();
     const stack = new Stack(app);
-    new AmplifyAuth(stack, 'test', { loginWith: { phoneNumber: true } });
+    new AmplifyAuth(stack, 'test', { loginWith: { phone: true } });
     const template = Template.fromStack(stack);
     template.hasResourceProperties('AWS::Cognito::UserPool', {
       UsernameAttributes: ['phone_number'],
@@ -131,7 +131,7 @@ void describe('Auth construct', () => {
   void it('creates phone login mechanism if settings is empty object', () => {
     const app = new App();
     const stack = new Stack(app);
-    new AmplifyAuth(stack, 'test', { loginWith: { phoneNumber: {} } });
+    new AmplifyAuth(stack, 'test', { loginWith: { phone: {} } });
     const template = Template.fromStack(stack);
     template.hasResourceProperties('AWS::Cognito::UserPool', {
       UsernameAttributes: ['phone_number'],
@@ -139,27 +139,42 @@ void describe('Auth construct', () => {
     });
   });
 
+  void it('throws error login settings do not include at least phone or email', () => {
+    const app = new App();
+    const stack = new Stack(app);
+    assert.throws(
+      () =>
+        new AmplifyAuth(stack, 'test', {
+          loginWith: {},
+        }),
+      {
+        message: 'At least one of email or phone must be enabled.',
+      }
+    );
+  });
+
   void it('creates email login mechanism with specific settings', () => {
     const app = new App();
     const stack = new Stack(app);
-    const customEmailVerificationMessage = 'custom email body {####}';
+    const emailBodyFunction = (code: string) => `custom email body ${code}`;
+    const expectedEmailMessage = 'custom email body {####}';
     const customEmailVerificationSubject = 'custom subject';
     new AmplifyAuth(stack, 'test', {
       loginWith: {
         email: {
-          verificationEmailBody: customEmailVerificationMessage,
-          verificationEmailStyle: VerificationEmailStyle.CODE,
+          verificationEmailBody: emailBodyFunction,
+          verificationEmailStyle: 'CODE',
           verificationEmailSubject: customEmailVerificationSubject,
         },
       },
     });
     const template = Template.fromStack(stack);
     template.hasResourceProperties('AWS::Cognito::UserPool', {
-      EmailVerificationMessage: customEmailVerificationMessage,
+      EmailVerificationMessage: expectedEmailMessage,
       EmailVerificationSubject: customEmailVerificationSubject,
       VerificationMessageTemplate: {
         DefaultEmailOption: 'CONFIRM_WITH_CODE',
-        EmailMessage: customEmailVerificationMessage,
+        EmailMessage: expectedEmailMessage,
         EmailSubject: customEmailVerificationSubject,
         SmsMessage: 'The verification code to your new account is {####}',
       },
@@ -169,65 +184,69 @@ void describe('Auth construct', () => {
   void it('creates email login mechanism with MFA', () => {
     const app = new App();
     const stack = new Stack(app);
-    const customEmailVerificationMessage = 'custom email body {####}';
+    const emailBodyFunction = (code: string) => `custom email body ${code}`;
+    const expectedEmailMessage = 'custom email body {####}';
     const customEmailVerificationSubject = 'custom subject';
-    const smsVerificationMessage = 'the verification code is {####}';
-    const smsMFAMessage = 'SMS MFA code is {####}';
+    const smsVerificationMessageFunction = (code: string) =>
+      `the verification code is ${code}`;
+    const expectedSMSVerificationMessage = 'the verification code is {####}';
+    const smsMFAMessageFunction = (code: string) => `SMS MFA code is ${code}`;
+    const expectedSMSMFAMessage = 'SMS MFA code is {####}';
     new AmplifyAuth(stack, 'test', {
       loginWith: {
         email: {
-          verificationEmailBody: customEmailVerificationMessage,
-          verificationEmailStyle: VerificationEmailStyle.CODE,
+          verificationEmailBody: emailBodyFunction,
+          verificationEmailStyle: 'CODE',
           verificationEmailSubject: customEmailVerificationSubject,
         },
-        phoneNumber: {
-          verificationMessage: smsVerificationMessage,
+        phone: {
+          verificationMessage: smsVerificationMessageFunction,
         },
       },
       multifactor: {
         enforcementType: 'OPTIONAL',
         sms: {
-          smsMessage: smsMFAMessage,
+          smsMessage: smsMFAMessageFunction,
         },
         totp: false,
       },
     });
     const template = Template.fromStack(stack);
     template.hasResourceProperties('AWS::Cognito::UserPool', {
-      EmailVerificationMessage: customEmailVerificationMessage,
+      EmailVerificationMessage: expectedEmailMessage,
       EmailVerificationSubject: customEmailVerificationSubject,
       VerificationMessageTemplate: {
         DefaultEmailOption: 'CONFIRM_WITH_CODE',
-        EmailMessage: customEmailVerificationMessage,
+        EmailMessage: expectedEmailMessage,
         EmailSubject: customEmailVerificationSubject,
-        SmsMessage: smsVerificationMessage,
+        SmsMessage: expectedSMSVerificationMessage,
       },
       MfaConfiguration: 'OPTIONAL',
       EnabledMfas: ['SMS_MFA'],
-      SmsAuthenticationMessage: smsMFAMessage,
-      SmsVerificationMessage: smsVerificationMessage,
+      SmsAuthenticationMessage: expectedSMSMFAMessage,
+      SmsVerificationMessage: expectedSMSVerificationMessage,
     });
   });
 
   void it('throws error if invalid email verification message for CODE', () => {
     const app = new App();
     const stack = new Stack(app);
-    const customEmailVerificationMessage = 'invalid message without code';
+    const emailBodyFunction = () => 'invalid message without code';
     const customEmailVerificationSubject = 'custom subject';
     assert.throws(
       () =>
         new AmplifyAuth(stack, 'test', {
           loginWith: {
             email: {
-              verificationEmailBody: customEmailVerificationMessage,
-              verificationEmailStyle: VerificationEmailStyle.CODE,
+              verificationEmailBody: emailBodyFunction,
+              verificationEmailStyle: 'CODE',
               verificationEmailSubject: customEmailVerificationSubject,
             },
           },
         }),
       {
         message:
-          "Invalid email settings. Property 'emailBody' must contain {####} as a placeholder for the verification code.",
+          "Invalid email settings. Property 'verificationEmailBody' must utilize the 'code' parameter at least once as a placeholder for the verification code.",
       }
     );
   });
@@ -235,22 +254,22 @@ void describe('Auth construct', () => {
   void it('throws error if invalid email verification message for LINK', () => {
     const app = new App();
     const stack = new Stack(app);
-    const customEmailVerificationMessage = 'invalid message without link';
+    const emailBodyFunction = () => 'invalid message without link';
     const customEmailVerificationSubject = 'custom subject';
     assert.throws(
       () =>
         new AmplifyAuth(stack, 'test', {
           loginWith: {
             email: {
-              verificationEmailBody: customEmailVerificationMessage,
-              verificationEmailStyle: VerificationEmailStyle.LINK,
+              verificationEmailBody: emailBodyFunction,
+              verificationEmailStyle: 'LINK',
               verificationEmailSubject: customEmailVerificationSubject,
             },
           },
         }),
       {
         message:
-          "Invalid email settings. Property 'emailBody' must contain {##Verify Email##} as a placeholder for the verification link.",
+          "Invalid email settings. Property 'verificationEmailBody' must utilize the 'link' parameter at least once as a placeholder for the verification link.",
       }
     );
   });
@@ -258,16 +277,16 @@ void describe('Auth construct', () => {
   void it('does not throw if valid email verification message for LINK', () => {
     const app = new App();
     const stack = new Stack(app);
-    const customEmailVerificationMessage =
-      'valid message {##Verify Email##} with link';
+    const emailBodyFunction = (link: string) =>
+      `valid message ${link} with link`;
     const customEmailVerificationSubject = 'custom subject';
     assert.doesNotThrow(
       () =>
         new AmplifyAuth(stack, 'test', {
           loginWith: {
             email: {
-              verificationEmailBody: customEmailVerificationMessage,
-              verificationEmailStyle: VerificationEmailStyle.LINK,
+              verificationEmailBody: emailBodyFunction,
+              verificationEmailStyle: 'LINK',
               verificationEmailSubject: customEmailVerificationSubject,
             },
           },
@@ -278,19 +297,19 @@ void describe('Auth construct', () => {
   void it('throws error if invalid sms verification message', () => {
     const app = new App();
     const stack = new Stack(app);
-    const customSMSVerificationMessage = 'invalid message without code';
+    const smsVerificationMessageFunction = () => 'invalid message without code';
     assert.throws(
       () =>
         new AmplifyAuth(stack, 'test', {
           loginWith: {
-            phoneNumber: {
-              verificationMessage: customSMSVerificationMessage,
+            phone: {
+              verificationMessage: smsVerificationMessageFunction,
             },
           },
         }),
       {
         message:
-          "Invalid phoneNumber settings. Property 'verificationMessage' must contain {####} as a placeholder for the verification code.",
+          "Invalid phone settings. Property 'verificationMessage' must utilize the 'code' parameter at least once as a placeholder for the verification code.",
       }
     );
   });
@@ -298,7 +317,7 @@ void describe('Auth construct', () => {
   void it('does not throw error if valid MFA message', () => {
     const app = new App();
     const stack = new Stack(app);
-    const validMFAMessage = 'valid MFA message with {####}';
+    const validMFAMessage = (code: string) => `valid MFA message with ${code}`;
     assert.doesNotThrow(
       () =>
         new AmplifyAuth(stack, 'test', {
@@ -319,7 +338,7 @@ void describe('Auth construct', () => {
   void it('throws error if invalid MFA message', () => {
     const app = new App();
     const stack = new Stack(app);
-    const invalidMFAMessage = 'invalid MFA message without code';
+    const invalidMFAMessage = () => 'invalid MFA message without code';
     assert.throws(
       () =>
         new AmplifyAuth(stack, 'test', {
@@ -336,7 +355,7 @@ void describe('Auth construct', () => {
         }),
       {
         message:
-          "Invalid MFA settings. Property 'smsMessage' must contain {####} as a placeholder for the verification code.",
+          "Invalid MFA settings. Property 'smsMessage' must utilize the 'code' parameter at least once as a placeholder for the verification code.",
       }
     );
   });
@@ -365,7 +384,7 @@ void describe('Auth construct', () => {
     const app = new App();
     const stack = new Stack(app);
     new AmplifyAuth(stack, 'test', {
-      loginWith: { phoneNumber: true, email: true },
+      loginWith: { phone: true, email: true },
       accountRecovery: AccountRecovery.EMAIL_AND_PHONE_WITHOUT_MFA,
     });
     const template = Template.fromStack(stack);
@@ -445,7 +464,7 @@ void describe('Auth construct', () => {
         authConstruct.node.findChild('IdentityPool') as CfnIdentityPool
       ).ref;
       const expectedWebClientId = (
-        authConstruct.node.findChild('UserPoolWebClient') as UserPoolClient
+        authConstruct.node.findChild('UserPoolAppClient') as UserPoolClient
       ).userPoolClientId;
       const expectedRegion = Stack.of(authConstruct).region;
 
@@ -599,7 +618,7 @@ void describe('Auth construct', () => {
     void it('sets account recovery settings to phone if phone is the only login type', () => {
       const app = new App();
       const stack = new Stack(app);
-      new AmplifyAuth(stack, 'test', { loginWith: { phoneNumber: true } });
+      new AmplifyAuth(stack, 'test', { loginWith: { phone: true } });
       const template = Template.fromStack(stack);
       template.hasResourceProperties('AWS::Cognito::UserPool', {
         AccountRecoverySetting: {
@@ -617,7 +636,7 @@ void describe('Auth construct', () => {
       const app = new App();
       const stack = new Stack(app);
       new AmplifyAuth(stack, 'test', {
-        loginWith: { phoneNumber: true, email: true },
+        loginWith: { phone: true, email: true },
       });
       const template = Template.fromStack(stack);
       template.hasResourceProperties('AWS::Cognito::UserPool', {
@@ -829,8 +848,7 @@ void describe('Auth construct', () => {
           externalProviders: {
             google: {
               clientId: googleClientId,
-              clientSecretValue:
-                SecretValue.unsafePlainText(googleClientSecret),
+              clientSecret: SecretValue.unsafePlainText(googleClientSecret),
             },
           },
         },
@@ -855,12 +873,11 @@ void describe('Auth construct', () => {
       const stack = new Stack(app);
       new AmplifyAuth(stack, 'test', {
         loginWith: {
-          phoneNumber: true,
+          phone: true,
           externalProviders: {
             google: {
               clientId: googleClientId,
-              clientSecretValue:
-                SecretValue.unsafePlainText(googleClientSecret),
+              clientSecret: SecretValue.unsafePlainText(googleClientSecret),
             },
           },
         },
@@ -914,7 +931,7 @@ void describe('Auth construct', () => {
       const stack = new Stack(app);
       new AmplifyAuth(stack, 'test', {
         loginWith: {
-          phoneNumber: true,
+          phone: true,
           externalProviders: {
             facebook: {
               clientId: facebookClientId,
@@ -974,7 +991,7 @@ void describe('Auth construct', () => {
       const stack = new Stack(app);
       new AmplifyAuth(stack, 'test', {
         loginWith: {
-          phoneNumber: true,
+          phone: true,
           externalProviders: {
             signInWithApple: {
               clientId: appleClientId,
@@ -1034,7 +1051,7 @@ void describe('Auth construct', () => {
       const stack = new Stack(app);
       new AmplifyAuth(stack, 'test', {
         loginWith: {
-          phoneNumber: true,
+          phone: true,
           externalProviders: {
             loginWithAmazon: {
               clientId: amazonClientId,
@@ -1089,7 +1106,7 @@ void describe('Auth construct', () => {
       const stack = new Stack(app);
       new AmplifyAuth(stack, 'test', {
         loginWith: {
-          phoneNumber: true,
+          phone: true,
           externalProviders: {
             oidc: {
               clientId: oidcClientId,
@@ -1142,7 +1159,7 @@ void describe('Auth construct', () => {
       const stack = new Stack(app);
       new AmplifyAuth(stack, 'test', {
         loginWith: {
-          phoneNumber: true,
+          phone: true,
           externalProviders: {
             saml: {
               name: samlProviderName,
@@ -1165,18 +1182,56 @@ void describe('Auth construct', () => {
       );
     });
 
+    void it('supports additional oauth settings', () => {
+      const app = new App();
+      const stack = new Stack(app);
+      new AmplifyAuth(stack, 'test', {
+        loginWith: {
+          email: true,
+          externalProviders: {
+            google: {
+              clientId: googleClientId,
+              clientSecret: SecretValue.unsafePlainText(googleClientSecret),
+            },
+            scopes: [OAuthScope.EMAIL, OAuthScope.PROFILE],
+            callbackUrls: ['http://localhost'],
+            logoutUrls: ['http://localhost'],
+          },
+        },
+      });
+      const template = Template.fromStack(stack);
+      template.hasResourceProperties('AWS::Cognito::UserPool', {
+        UsernameAttributes: ['email'],
+        AutoVerifiedAttributes: ['email'],
+      });
+      template.hasResourceProperties(
+        'AWS::Cognito::UserPoolIdentityProvider',
+        ExpectedGoogleIDPProperties
+      );
+      template.hasResourceProperties('AWS::Cognito::IdentityPool', {
+        SupportedLoginProviders: {
+          'accounts.google.com': googleClientId,
+        },
+      });
+      template.hasResourceProperties('AWS::Cognito::UserPoolClient', {
+        PreventUserExistenceErrors: 'ENABLED',
+        CallbackURLs: ['http://localhost'],
+        LogoutURLs: ['http://localhost'],
+        AllowedOAuthScopes: ['email', 'profile', 'openid'],
+      });
+    });
+
     void it('supports all idps and login methods', () => {
       const app = new App();
       const stack = new Stack(app);
       new AmplifyAuth(stack, 'test', {
         loginWith: {
           email: true,
-          phoneNumber: true,
+          phone: true,
           externalProviders: {
             google: {
               clientId: googleClientId,
-              clientSecretValue:
-                SecretValue.unsafePlainText(googleClientSecret),
+              clientSecret: SecretValue.unsafePlainText(googleClientSecret),
             },
             facebook: {
               clientId: facebookClientId,