+++ packages/backend-auth/src/translate_auth_props.test.ts
@@ -16,8 +16,8 @@ import assert from 'node:assert';
 import { translateToAuthConstructLoginWith } from './translate_auth_props.js';
 import { BranchBackendIdentifier } from '@aws-amplify/platform-core';
 
-const phoneNumber: PhoneNumberLogin = {
-  verificationMessage: 'text{####}text2',
+const phone: PhoneNumberLogin = {
+  verificationMessage: (code: string) => `text${code}text2`,
 };
 const googleClientId = 'googleId';
 const googleClientSecret = 'googleSecret';
@@ -57,11 +57,11 @@ void describe('translateToAuthConstructLoginWith', () => {
 
   void it('translates with external providers', () => {
     const loginWith: AuthLoginWithFactoryProps = {
-      phoneNumber,
+      phone,
       externalProviders: {
         google: {
           clientId: new TestBackendSecret(googleClientId),
-          clientSecretValue: new TestBackendSecret(googleClientSecret),
+          clientSecret: new TestBackendSecret(googleClientSecret),
         },
         facebook: {
           clientId: new TestBackendSecret(facebookClientId),
@@ -92,11 +92,11 @@ void describe('translateToAuthConstructLoginWith', () => {
     );
 
     const expected: BasicLoginOptions & ExternalProviderProps = {
-      phoneNumber,
+      phone,
       externalProviders: {
         google: {
           clientId: googleClientId,
-          clientSecretValue: SecretValue.unsafePlainText(googleClientSecret),
+          clientSecret: SecretValue.unsafePlainText(googleClientSecret),
         },
         facebook: {
           clientId: facebookClientId,
@@ -125,7 +125,7 @@ void describe('translateToAuthConstructLoginWith', () => {
 
   void it('translates with only a general provider attribute', () => {
     const loginWith: AuthLoginWithFactoryProps = {
-      phoneNumber,
+      phone,
       externalProviders: {
         callbackUrls: callbackUrls,
       },
@@ -137,7 +137,7 @@ void describe('translateToAuthConstructLoginWith', () => {
     );
 
     const expected: BasicLoginOptions & ExternalProviderProps = {
-      phoneNumber,
+      phone,
       externalProviders: {
         callbackUrls: callbackUrls,
       },
@@ -147,7 +147,7 @@ void describe('translateToAuthConstructLoginWith', () => {
 
   void it('translates without external providers', () => {
     const loginWith: AuthLoginWithFactoryProps = {
-      phoneNumber,
+      phone,
     };
 
     const translated = translateToAuthConstructLoginWith(
@@ -156,7 +156,7 @@ void describe('translateToAuthConstructLoginWith', () => {
     );
 
     const expected: BasicLoginOptions & ExternalProviderProps = {
-      phoneNumber,
+      phone,
     };
     assert.deepStrictEqual(translated, expected);
   });